+++ src/basilisp_blender/nrepl.py
@@ -8,7 +8,6 @@ from basilisp.lang import keyword as kw
 from basilisp.lang import map as lmap
 from basilisp.lang.util import munge
 
-
 def server_thread_async_start(host="127.0.0.1", port=0, nrepl_port_filepath=None):
     """Start an nREPL server on the specified `host` and `port` on a
     separate thread.
@@ -27,8 +26,8 @@ def server_thread_async_start(host="127.0.0.1", port=0, nrepl_port_filepath=None
     assert '"' not in host
     assert port >= 0
 
-    nrepl_server_mod = importlib.import_module(munge("basilisp-blender.nrepl-server"))
-    ret = nrepl_server_mod.start_server__BANG__(
+    nrepl_server_mod = importlib.import_module(munge("basilisp-nrepl-async.nrepl-server"))
+    ret = nrepl_server_mod.server_start__BANG__(
         lmap.map(
             {
                 kw.keyword("async?"): True,